+++ src/parsers/tuparser.h
@@ -10,8 +10,8 @@ public:
     explicit TUParser(QObject *parent = 0);
     bool parseFile(std::istream& pcInputStream, ComSequence* pcSequence);
 protected:
-    size_t xReadTU(const std::vector<uchar>& vPCInfo, size_t s, ComSequence* sequence, ComCU* pcTU);
-    size_t xReadTUHelper(const std::vector<uchar>& vPCInfo, size_t s, ComSequence* sequence, ComTU* pcTU);
+    size_t xReadTU(const std::vector<int>& vPCInfo, size_t s, ComSequence* sequence, ComCU* pcTU);
+    size_t xReadTUHelper(const std::vector<int>& vPCInfo, size_t s, ComSequence* sequence, ComTU* pcTU);
 };
 
 #endif // TUPARSER_H
